+++ src/main/java/com/ject/vs/chat/adapter/web/dto/MessagePageResponse.java
@@ -1,6 +1,6 @@
 package com.ject.vs.chat.adapter.web.dto;
 
-import com.ject.vs.chat.port.in.MessagePageResult;
+import com.ject.vs.chat.port.in.dto.MessagePageResult;
 
 import java.util.List;
