+++ src/main/java/com/ject/vs/chat/port/ChatCommandService.java
@@ -7,9 +7,9 @@ import com.ject.vs.chat.domain.ChatRoomUnreadRepository;
 import com.ject.vs.chat.exception.ChatForbiddenException;
 import com.ject.vs.chat.exception.InvalidMessageException;
 import com.ject.vs.chat.port.in.ChatCommandUseCase;
-import com.ject.vs.chat.port.in.MarkAsReadCommand;
-import com.ject.vs.chat.port.in.MessageResult;
-import com.ject.vs.chat.port.in.SendMessageCommand;
+import com.ject.vs.chat.port.in.dto.MarkAsReadCommand;
+import com.ject.vs.chat.port.in.dto.MessageResult;
+import com.ject.vs.chat.port.in.dto.SendMessageCommand;
 import com.ject.vs.domain.User;
 import com.ject.vs.repository.UserRepository;
 import com.ject.vs.vote.domain.VoteParticipationRepository;
