+++ src/main/java/com/ject/vs/chat/port/in/ChatQueryUseCase.java
@@ -1,5 +1,11 @@
 package com.ject.vs.chat.port.in;
 
+import com.ject.vs.chat.port.in.dto.ChatListItemResult;
+import com.ject.vs.chat.port.in.dto.ChatRoomResult;
+import com.ject.vs.chat.port.in.dto.GaugeResult;
+import com.ject.vs.chat.port.in.dto.MessagePageResult;
+import com.ject.vs.chat.port.in.dto.VoteStatus;
+
 import java.util.List;
 
 public interface ChatQueryUseCase {
