+++ src/main/java/com/ject/vs/chat/port/in/dto/GaugeResult.java
@@ -1,3 +1,3 @@
-package com.ject.vs.chat.port.in;
+package com.ject.vs.chat.port.in.dto;
 
 public record GaugeResult(int optionARatio, int optionBRatio, int participantCount) {}
