+++ src/main/java/com/ject/vs/chat/port/in/dto/MarkAsReadCommand.java
@@ -1,3 +1,3 @@
-package com.ject.vs.chat.port.in;
+package com.ject.vs.chat.port.in.dto;
 
 public record MarkAsReadCommand(Long voteId, Long userId, Long lastReadMessageId) {}
