+++ src/main/java/com/ject/vs/chat/port/in/dto/SendMessageCommand.java
@@ -1,3 +1,3 @@
-package com.ject.vs.chat.port.in;
+package com.ject.vs.chat.port.in.dto;
 
 public record SendMessageCommand(Long voteId, Long senderId, String content) {}
