+++ src/main/java/com/ject/vs/chat/port/in/dto/VoteStatus.java
@@ -1,4 +1,4 @@
-package com.ject.vs.chat.port.in;
+package com.ject.vs.chat.port.in.dto;
 
 public enum VoteStatus {
     ONGOING, ENDED
